error: 출력 중 문제 발생
