+++ README.md
@@ -17,7 +17,7 @@
 
 ### 3. 출력부
 
-- 로또 구입 개수 출력
+- 로또 구입 개수 출력 **→ 문제 발생**
 - 발행 로또 출력
 - 당첨 내역 출력
 - 수익률 출력
