+++ src/main/java/lotto/Application.java
@@ -1,7 +1,14 @@
 package lotto;
 
+import lotto.view.InputMoney;
+import lotto.view.OutputView;
+
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        InputMoney money = new InputMoney();
+        OutputView output = new OutputView();
+
+        money.getMoney();
+        output.showLottoCount();
     }
 }
