+++ src/main/java/lotto/model/Lotto.java
@@ -1,7 +1,5 @@
 package lotto.model;
 
-import lotto.view.InputMoney;
-
 import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.ArrayList;
@@ -19,16 +17,6 @@ public class Lotto {
         if (numbers.size() != 6) {
             throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
         }
-    }
-
-    // 로또 구입 개수 연산
-    public int setPurchaseLotto() {
-        InputMoney tmpMoney = new InputMoney();
-
-        int money = tmpMoney.getMoney();
-        int lotto = money % 1000;
-
-        return lotto;
     }
 
     // 랜덤한 로또 발행 연산
